requirements.txt
ultralytics
cvzone
opencv-python
telegram
requests
cryptography
python-dotenv
pytest
flask
flask-cors
python-dotenv
pymongo
flask-bcrypt
flask_mail